Adding support for DESTDIR to Makefile, and a new 'uninstall' target.
As per the GNU Makefile Conventions, DESTDIR is a value that can be
prepended to the installation target paths in order to support custom
installation paths, as well as staged installs (where files are placed
in an intermediate path first). It is also used in the Debian packaging
proces (cf. https://www.gnu.org/prep/standards/html_node/DESTDIR.html)

The 'uninstall' target can be used to remove the program, should the
user wish to do so.