Fixing Makefile to allow installation to nonexistent paths.
Due to a bug in my previous commit (ee56e68), the Makefile could only be
used to install the program and manpage to directories that already exist.

This makes it impossible to use for building into a sandbox environment, or
anywhere that doesn't already have /usr/bin and /usr/share/man/man6. In
particular, this poses problems for packaging.