Always output aimation to terminal
In case of commands with stdout redirection, Gti used to write the
animation into the redirected stdout. This means that the user does
not get to see the nice little animation. Clearly, this isn't what
we want.

Rewrite the code to open /dev/tty instead, and write to that. For
systems lacking /dev/tty, we gracefully fall back to stdout.